moving polyfills out to another file, removing extraneous entries, updated imports to use exports instead
