+++ packages/spark-core/spark-core-polyfills.js
@@ -1,0 +1,6 @@
+// Polyfills
+import './utilities/polyfills/ArrayFrom';
+import './utilities/polyfills/StringIncludes';
+import './utilities/polyfills/ArrayFind';
+import './utilities/polyfills/NodeListForEach';
+import './utilities/polyfills/classListSVG';
