+++ src/assets/drizzle/scripts/pagination/default.js
@@ -12,7 +12,7 @@
  * data-sprk-pagination="next" and data-sprk-pagination="prev"
  * These identify the previous and next links in the pagination components.
  */
-import getElements from '../../../../../packages/spark-core/es5/getElements';
+import { getElements } from '../../../../../packages/spark-core';
 
 // Add aria-current attr to new page link, remove from old
 const setAriaCurrent = (newPageLink, oldPageLink) => {
