+++ src/assets/toolkit/scripts/toolkit.js
@@ -1,4 +1,5 @@
 import sparkCore from '../../../../packages/spark-core/es5/sparkCore';
+import '../../../../packages/spark-core/es5/sparkCorePolyfills';
 import highlightBoard from '../../../../packages/spark-extras/components/highlight-board/es5/highlightBoard';
 
 sparkCore({
